Extracts version detection into separate file.
