+++ packages/cosmo/package.py
@@ -3,39 +3,9 @@
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
-import subprocess, re, itertools, os
+import itertools, os
 from spack import *
-
-
-def get_releases(repo):
-    git_obj = subprocess.run(["git", "ls-remote", "--refs", repo],
-                             capture_output=True)
-    if git_obj.returncode != 0:
-        print("\nWarning: no access to {:s} => not fetching versions\n".format(
-            repo))
-        return []
-    else:
-        git_tags = [
-            re.match('refs/tags/(.*)', x.decode('utf-8')).group(1)
-            for x in git_obj.stdout.split()
-            if re.match('refs/tags/(.*)', x.decode('utf-8'))
-        ]
-        return git_tags
-
-
-def set_versions(repo, reg_filter=None):
-
-    def filterfn(repo_tag):
-        return re.match(reg_filter, repo_tag) != None
-
-    tags = get_releases(repo)
-    if tags:
-        if reg_filter:
-            tags = list(filter(filterfn, tags))
-
-        for tag in tags:
-            version(tag, git=repo, tag=tag, get_full_repo=True)
-
+from version_detection import set_versions
 
 class Cosmo(MakefilePackage):
     """COSMO: Numerical Weather Prediction Model. Needs access to private GitHub."""
@@ -63,7 +33,7 @@ class Cosmo(MakefilePackage):
     patch('patches/5.07.mch1.0.p4/patch.Makefile', when='@5.07.mch1.0.p4')
     patch('patches/5.07.mch1.0.p4/patch.Makefile', when='@5.07.mch1.0.p5')
 
-    set_versions(apngit, reg_filter='.*mch.*')
+    set_versions(apngit, regex_filter='.*mch.*')
     set_versions(c2smgit)
 
     depends_on('netcdf-fortran', type=('build', 'link'))
